requirements.txt
streamlit==1.32.0
ultralytics==8.1.2
easyocr==1.7.2
pyzbar==0.1.9
pandas==2.2.0
opencv-python==4.11.0.86
pillow==10.2.0 